requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
gTTS>=2.3.2
pydub>=0.25.1
torch>=2.0.0
librosa>=0.10.1
soundfile>=0.12.1
requests>=2.31.0
ffmpeg-python>=0.2.0
